seperated into init and initplus
